made many changes
Some of which are recommended by ChatGPT.

Added a global error handler.

Removed the old error and request loggers and added new ones. Idk if I actually need this, but whatever.

I improved articleReader, added a try catch for the reader, and made a custom error to throw if the filePath isn't a string.

In articleController and categoryController, i added some .catch(), logError(err), and logRequest(req) where it was needed.
Also, i put categoryObject between brackets so now i don't need to use findOne to get the category that i just created.

Also made a few other small changed that are not really worth mentioning.